Refactor cudnn convolution
